pyproject.toml
[project]
name = "chaos"
version = "0.1.0"
description = "Collaborative Hierarchical Agents for Open Sensemaking - A multi-agent LLM system for sensemaking"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Miklos Havlik", email = "s240069@dtu.dk" }, { name = "Peter Jahola", email = "s233734@dtu.dk" }]
keywords = ["llm", "agents", "multi-agent", "sensemaking"]

dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "numpy>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
chaos = "chaos.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=chaos"
